human: modify

+++ gmm/grouped_matmul/op_kernel/arch35/weight_quant_basic_block/block/grouped_matmul_scheduler_n_resplit.h
@@ -16,7 +16,7 @@ public:
         uint64_t secondTailBlockSize;
         uint64_t coreNum;
         uint64_t cubeNumBlocksN;
-        uint64_t baseM;
+        int32_t baseM;
         uint64_t nSize;
     };
